# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webmail API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface InboxItem {
  id: string;
  from: string;
  subject: string;
  date: string;
  isRead: boolean;
}

export interface InboxResponse {
  email: string;
  messages: InboxItem[];
}

export interface EmailMessage {
  id: string;
  from: string;
  subject: string;
  date: string;
  html: string;
}

export interface DomainItem {
  id: number;
  domain: string;
  /** @nullable */
  label?: string | null;
  createdAt: string;
}

export interface DomainListResponse {
  domains: DomainItem[];
}

export interface DomainInput {
  /** The domain name (e.g. mail.yoursite.com) */
  domain: string;
  /**
     * Optional friendly label for the domain
     * @nullable
     */
  label?: string | null;
}

export type GetInboxParams = {
/**
 * Full email address (e.g. user@yourdomain.com)
 */
email: string;
};

export type GetEmailParams = {
/**
 * Full email address
 */
email: string;
/**
 * Message ID
 */
id: string;
};