# lib/api-zod/src/generated/types/domainInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webmail API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DomainInput {
  /** The domain name (e.g. mail.yoursite.com) */
  domain: string;
  /**
     * Optional friendly label for the domain
     * @nullable
     */
  label?: string | null;
}