# lib/api-zod/src/generated/types/domainItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webmail API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DomainItem {
  id: number;
  domain: string;
  /** @nullable */
  label?: string | null;
  createdAt: string;
}