# lib/api-zod/src/generated/types/domainListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webmail API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DomainItem } from './domainItem';

export interface DomainListResponse {
  domains: DomainItem[];
}