# lib/api-zod/src/generated/types/inboxItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webmail API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InboxItem {
  id: string;
  from: string;
  subject: string;
  date: string;
  isRead: boolean;
}