# lib/api-zod/src/generated/types/inboxResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webmail API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InboxItem } from './inboxItem';

export interface InboxResponse {
  email: string;
  messages: InboxItem[];
}